Updated DAI MOJO NiFi Deployment based on Nick & Edgar's Feedback
- In Maven pom.xml, added mojo2-runtime-api and mojo2-runtime-impl for Maven to manage dependencies
- In mojo-nifi/README.md, updated split command to use numeric suffix and additional suffix
- In files that had Apache License 2.0 comments, removed them since one single License is autogenerated
when user builds project with mvn clean install
- In pom.xml, updated all constant nifi versions with Maven nifi.version variable
- In Java both java programs, declared more specific exception types
- In Java test program, updated setup method to indicate that certain Exceptions may be thrown
- In Java test program, added static to randFloat
- In Java test program, removed unused Random r variable
- In Java test program, added helper generateDefaultHydraulicTestData(numRecords) method to encapsulate null
- In NiFi Java test programs, it looks like they mark their tests as public with name convention test<Description()>
- In Java ExecuteDaiMojoScoringPipeline, changed _properties to propertiesBuilder and did similar for other variables
that started with underscores
- Also added comments above methods for further explaining method if it was needed
- For code that was self-documenting, removed comments
- For convertMojoFrameMetaToRecordSchema() method and convertMojoTypeToRecordType(MojoColumn.Type mojoType), made sure
to deal with null based on suggestion, so now the logic should be correct. As you can see also moved the switch statement
to convertMojoTypeToRecordType(MojoColumn.Type mojoType)
- Removed customValidate method since NiFi comes with plenty of StandardValidators. Used a NiFi validator to verify that the
file for the pipeline mojo filepath property exists
- In most places, changed getLogger() to logger
- At the last catch(...) in onTrigger() method, added more specific Exceptions to catch and removed final keyword from line 327
- For recordMap, separated it into 2 lines for easier readability: initialize dataType, then initialize recordMap with record
and dataType
- For switch case statement that was checking mojoFrameColDataType and then converting mojoColumnData to appropriate dataType,

moved it to getMojoFrameDataEntry(...)
