+++ mojo-nifi/nifi-nar-bundles/nifi-daimojo-record-bundle/nifi-daimojo-record-processors/src/main/resources/META-INF/services/org.apache.nifi.processor.Processor
@@ -1,0 +1,1 @@
+org.apache.nifi.processors.daimojo.record.ExecuteDaiMojoScoringPipeline
